Make it work with 32-bits system

+++ src/three-address-code-simulator/sim.cpp
@@ -111,7 +111,7 @@ string toString(string src) {
     case TypeName::Double:
         return to_string(t.value.d);
     case TypeName::Pointer:
-        return to_string((unsigned long long)t.value.ptr);
+        return to_string((unsigned long)t.value.ptr);
     default:
         return "";
     }
@@ -669,7 +669,7 @@ void executeBuildinFunc(string func, string des){
     if(func == "f_malloc"){
         Var t = argStack.top();
         argStack.pop();
-        setVal(des, TypeName::Pointer, to_string((unsigned long long)malloc(t.value.uint8)));
+        setVal(des, TypeName::Pointer, to_string((unsigned long)malloc(t.value.uint8)));
         return;
     }
     if(func == "f_free"){
@@ -685,7 +685,7 @@ void call_exe(string func, size_t argc){
         backupStack.push(symbol_table[string("p") + to_string(i - 1)]);
     }
     argSizeStack.push(argc);
-    pc = (unsigned long long)symbol_table[func].value.ptr;
+    pc = (unsigned long)symbol_table[func].value.ptr;
     for (size_t i = argc; i > 0; --i) {
         string var_name = string("p") + to_string(i - 1);
         symbol_table[var_name] = argStack.top();
@@ -732,7 +732,7 @@ void execute(int pc_l) {
         symbol_table[tmp].value.int4 = 0;
         s_op2(tmp, get<1>(t_ins), get<2>(t_ins), get<3>(t_ins));
         if (symbol_table["tmp"].value.int4) {
-            pc = (unsigned long long)symbol_table[get<4>(t_ins)].value.ptr;
+            pc = (unsigned long)symbol_table[get<4>(t_ins)].value.ptr;
         }
         return;
     }
@@ -756,7 +756,7 @@ void execute(int pc_l) {
         // return
         pc = pcStack.top();
         // pc == -1 mean that the program has exit the main function.
-        if(pc == (unsigned long long) -1)
+        if(pc == (unsigned long) -1)
             return;
         pcStack.pop();
         if (get<1>(t_ins) != "") {
@@ -878,13 +878,13 @@ int main(int argc, char **argv) {
     initialize(is);
 
 
-    pc = (unsigned long long)symbol_table["f_main"].value.ptr;
+    pc = (unsigned long)symbol_table["f_main"].value.ptr;
     pcStack.push(-1);
     argSizeStack.push(0);
 #ifdef DEBUG
     cout << "BEGIN PROGRAM" << endl;
 #endif
-    while (pc != (unsigned long long)-1) {
+    while (pc != (unsigned long)-1) {
         if(debugMode)
             debugFunc();
         execute(pc);
